Reorganize parser exceptions
