+++ src/php/Compiler/Emitter/EmitterInterface.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Phel\Compiler\Emitter;
 
 use Phel\Compiler\Analyzer\Ast\AbstractNode;
-use Phel\Exceptions\CompiledCodeIsMalformedException;
-use Phel\Exceptions\FileException;
+use Phel\Compiler\Emitter\Exceptions\CompiledCodeIsMalformedException;
+use Phel\Compiler\Emitter\Exceptions\FileException;
 
 interface EmitterInterface
 {
