+++ src/php/Compiler/Emitter/Exceptions/CompiledCodeIsMalformedException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Exceptions;
+namespace Phel\Compiler\Emitter\Exceptions;
 
 use RuntimeException;
 use Throwable;
