+++ src/php/Compiler/Emitter/Exceptions/FileException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Exceptions;
+namespace Phel\Compiler\Emitter\Exceptions;
 
 use RuntimeException;
