+++ src/php/Compiler/EvalCompiler.php
@@ -8,17 +8,17 @@ use Phel\Compiler\Analyzer\AnalyzerInterface;
 use Phel\Compiler\Analyzer\Environment\NodeEnvironment;
 use Phel\Compiler\Analyzer\Environment\NodeEnvironmentInterface;
 use Phel\Compiler\Analyzer\Exceptions\AnalyzerException;
+use Phel\Compiler\Analyzer\Exceptions\CompilerException;
 use Phel\Compiler\Emitter\EmitterInterface;
+use Phel\Compiler\Emitter\Exceptions\CompiledCodeIsMalformedException;
+use Phel\Compiler\Emitter\Exceptions\FileException;
 use Phel\Compiler\Lexer\LexerInterface;
+use Phel\Compiler\Parser\Exceptions\AbstractParserException;
+use Phel\Compiler\Parser\Exceptions\UnfinishedParserException;
 use Phel\Compiler\Parser\ParserInterface;
 use Phel\Compiler\Parser\ParserNode\TriviaNodeInterface;
 use Phel\Compiler\Parser\ReadModel\ReaderResult;
 use Phel\Compiler\Reader\ReaderInterface;
-use Phel\Exceptions\CompiledCodeIsMalformedException;
-use Phel\Exceptions\CompilerException;
-use Phel\Exceptions\FileException;
-use Phel\Exceptions\Parser\UnfinishedParserException;
-use Phel\Exceptions\ParserException;
 use Phel\Exceptions\ReaderException;
 
 final class EvalCompiler implements EvalCompilerInterface
@@ -68,7 +68,7 @@ final class EvalCompiler implements EvalCompilerInterface
             return $this->evalNode($readerResult);
         } catch (UnfinishedParserException $e) {
             throw $e;
-        } catch (ParserException|ReaderException $e) {
+        } catch (AbstractParserException|ReaderException $e) {
             throw new CompilerException($e, $e->getCodeSnippet());
         }
     }
