+++ src/php/Compiler/Parser/Exceptions/StringParserException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phel\Compiler\Parser\Exceptions;
+
+use RuntimeException;
+
+final class StringParserException extends RuntimeException
+{
+}
