+++ src/php/Compiler/Parser/Exceptions/UnfinishedParserException.php
@@ -2,13 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Phel\Exceptions\Parser;
+namespace Phel\Compiler\Parser\Exceptions;
 
 use Phel\Compiler\Lexer\Token;
 use Phel\Compiler\Parser\ReadModel\CodeSnippet;
-use Phel\Exceptions\ParserException;
 
-class UnfinishedParserException extends ParserException
+class UnfinishedParserException extends AbstractParserException
 {
     public static function forSnippet(CodeSnippet $snippet, Token $token, string $message): self
     {
