+++ src/php/Compiler/Parser/ExpressionParser/StringParser.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Phel\Compiler\Parser\ExpressionParser;
 
 use Phel\Compiler\Lexer\Token;
+use Phel\Compiler\Parser\Exceptions\StringParserException;
 use Phel\Compiler\Parser\ParserNode\StringNode;
-use Phel\Exceptions\StringParserException;
 
 final class StringParser
 {
