--- src/php/Exceptions/EmitterException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Phel\Exceptions;
-
-use Exception;
-
-final class EmitterException extends Exception
-{
-}
